Merge Two Arrays Together
Requires an if block with 4 conditions. Two of which compare the first
element of the left array and right array. The other two are edge cases
where if one array doesn't have any elements, push the elements in the
other array. For all conditions, the items are pushed to a 'merge array'
and the corresponding array shifts by one.